requirements.txt
streamlit
opencv-contrib-python-headless
numpy
Pillow
pytz